reorganize classifcation by calling external LLMs
